refactor: refactor and optimize unistyles web
