--- src/web/listener/index.ts
@@ -1,1 +1,0 @@
-export { UnistylesListener } from './listener'
